feat(V3X): robust NFO/MediaInfo handling and upload retry
The MediaInfo/NFO helpers (_patch_mi_filename, _get_mediainfo_text,
_get_or_generate_nfo) move from C411 into the shared French mixin. V3X
now embeds the release NFO in the torrent, falls back to MEDIAINFO.txt /
BDInfo / a generated scene NFO, patches the Complete name line on
generated NFOs (originals ship untouched), and retries the upload once
with a longer timeout on network errors.
