+++ src/trackers/C411.py
@@ -25,7 +25,6 @@ from unidecode import unidecode
 
 from src.console import console
 from src.get_desc import DescriptionBuilder
-from src.nfo_generator import SceneNfoGenerator
 from src.tmdb import TmdbManager
 from src.trackers.COMMON import COMMON
 from src.trackers.FRENCH import FrenchTrackerMixin
@@ -1139,79 +1138,6 @@ class C411(FrenchTrackerMixin):
 
         return "\n".join(parts)
 
-    @staticmethod
-    def _patch_mi_filename(mi_text: str, new_name: str) -> str:
-        """Replace the ‘Complete name’ value in MediaInfo text with *new_name*.
-
-        C411’s API validates that the filename inside the uploaded MediaInfo
-        matches the release name.  Since we rename releases (language tags,
-        -NOTAG label, French title…) the original filename no longer matches.
-        This patches the ‘Complete name’ line while preserving the file extension.
-        """
-        if not mi_text or not new_name:
-            return mi_text
-
-        def _replace_complete_name(match: re.Match[str]) -> str:
-            prefix = match.group(1)  # "Complete name    : "
-            old_value = match.group(2)
-            ext_match = re.search(r"(\.[a-zA-Z0-9]{2,4})$", old_value)
-            ext = ext_match.group(1) if ext_match else ""
-            return f"{prefix}{new_name}{ext}"
-
-        return re.sub(
-            r"^(Complete name\s*:\s*)(.+)$",
-            _replace_complete_name,
-            mi_text,
-            count=1,
-            flags=re.MULTILINE,
-        )
-
-    async def _get_mediainfo_text(self, meta: Meta) -> str:
-        """Read MediaInfo text from temp files.
-
-        The ‘Complete name’ line is patched to match the C411-generated
-        release name so that the API filename-consistency check passes.
-        """
-        base = os.path.join(meta.get("base_dir", ""), "tmp", meta.get("uuid", ""))
-        content = ""
-
-        # Prefer clean-path, then standard mediainfo
-        for fname in ("MEDIAINFO_CLEANPATH.txt", "MEDIAINFO.txt"):
-            fpath = os.path.join(base, fname)
-            if os.path.exists(fpath):
-                async with aiofiles.open(fpath, encoding="utf-8") as f:
-                    content = await f.read()
-                    if content.strip():
-                        break
-                    content = ""
-
-        # BDInfo for disc releases
-        if not content and meta.get("bdinfo") is not None:
-            bd_path = os.path.join(base, "BD_SUMMARY_00.txt")
-            if os.path.exists(bd_path):
-                async with aiofiles.open(bd_path, encoding="utf-8") as f:
-                    return await f.read()
-
-        # Fallback: use in-memory mediainfo from prep
-        if not content:
-            fallback = str(meta.get("mediainfo_text") or "").strip()
-            if fallback:
-                content = fallback
-
-        if not content:
-            return ""
-
-        # Patch “Complete name” to match the tracker-generated release name
-        try:
-            name_result = await self.get_name(meta)
-            tracker_release_name = name_result.get("name", "") if isinstance(name_result, dict) else str(name_result)
-            if tracker_release_name:
-                content = self._patch_mi_filename(content, tracker_release_name)
-        except Exception:
-            pass  # If naming fails, return unpatched MI
-
-        return content
-
     async def _build_tmdb_data(self, meta: Meta) -> Union[str, None]:
         """Build tmdbData JSON string for C411.
 
@@ -1363,34 +1289,6 @@ class C411(FrenchTrackerMixin):
     # ──────────────────────────────────────────────────────────
     #  NFO generation
     # ──────────────────────────────────────────────────────────
-
-    async def _get_or_generate_nfo(self, meta: Meta) -> Union[str, None]:
-        """Generate a MediaInfo-based NFO for the upload.
-
-        C411 requires an NFO file for every upload.
-        Either an NFO file generated by MediaInfo,
-        or an NFO file included with the release.
-        """
-        nfo_files = self._get_nfo_files(meta)
-        if nfo_files:
-            return nfo_files[0]
-        else:
-            return await self._get_or_generate_mediainfo_as_nfo(meta)
-
-    async def _get_or_generate_mediainfo_as_nfo(self, meta: Meta) -> Union[str, None]:
-        """Sub-function of _get_or_generate_nfo to get MI file if exists
-        Else, generate a NFO
-        """
-        mi_dir = os.path.join(meta.get("base_dir", ""), "tmp", meta.get("uuid", ""))
-        mi_clean = os.path.join(mi_dir, "MEDIAINFO_CLEANPATH.txt")
-        mi = os.path.join(mi_dir, "MEDIAINFO.txt")
-        if os.path.isfile(mi_clean):
-            return mi_clean
-        elif os.path.isfile(mi):
-            return mi
-        else:
-            nfo_gen = SceneNfoGenerator(self.config)
-            return await nfo_gen.generate_nfo(meta, self.tracker)
 
     # ──────────────────────────────────────────────────────────
     #  Upload / Search interface
